+++ CHANGELOG.md
@@ -3,8 +3,13 @@
 ## Buffomat Classic and TBC 2021.6.4
 
 - Changed task list behaviour, code cleanup
-- Changed task wording, important data goes first 
+- Changed task wording, important data goes first
 - TBC: Prevent dismount in flight when attempting to cast, OUCH!
+- Reminder to unequip carrot now works for riding crop and (not in the game yet)
+  druid flight speed trinket
+- Fixed reminders to unequip AD trinket, added new Naxx trinket versions, added
+  new zones: Auchenai and Karazhan to the list of places
+  where undead can be found.
 
 ## Buffomat Classic and TBC 2021.6.3